fixes and arguments

+++ conftest.py
@@ -7,13 +7,12 @@ import shutil
 from tsml_eval.experiments import experiments
 from tsml_eval.testing.test_utils import _TEST_OUTPUT_PATH
 
+KEEP_PYTEST_OUTPUT = True
+
 
 def pytest_sessionfinish(session, exitstatus):
-    """
-    Called after test run is finished, right before returning the exit status to
-    the system.
-    """
-    if not hasattr(session.config, "workerinput"):
+    """Call after test run is finished, before returning the exit status to system."""
+    if not hasattr(session.config, "workerinput") and not KEEP_PYTEST_OUTPUT:
         shutil.rmtree(_TEST_OUTPUT_PATH)
 
 
@@ -26,8 +25,16 @@ def pytest_addoption(parser):
         help="Set the time interval in seconds for recording memory usage "
         "(default: %(default)s).",
     )
+    parser.addoption(
+        "--keepoutput",
+        action="store_true",
+        help="Keep the unit test output folder after running pytest"
+        " (default: %(default)s).",
+    )
 
 
 def pytest_configure(config):
     """Pytest configuration preamble."""
     experiments.MEMRECORD_INTERVAL = config.getoption("--meminterval")
+    global KEEP_PYTEST_OUTPUT
+    KEEP_PYTEST_OUTPUT = config.getoption("--keepoutput")
